move graphics test from pegasus to lib
