+++ lib/test/fixtures/images/flag_sphere.png
