+++ lib/test/fixtures/images/hour-of-code-logo.png
