+++ lib/test/fixtures/images/kids4_2x.jpg
